% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNMFgenes}
\alias{getNMFgenes}
\title{Get list of genes for each NMF program}
\usage{
getNMFgenes(nmf.res, method = 0.5, max.genes = 50)
}
\arguments{
\item{nmf.res}{A list of NMF models obtained from `multiNMF`}

\item{method}{Parameter passed to `NMF::extractFeatures` to obtain top genes
for each program. When 'method' is a number between 0 and 1, it indicates
the minimum relative basis contribution above which the feature is
selected, i.e. how specific is a gene for a given program.}

\item{max.genes}{Max number of genes for each programs}
}
\value{
Returns a list of top genes for each gene program
}
\description{
Run it over a list of NMF models obtained using `multiNMF`
}
\examples{
library(Seurat)
data(sampleObj)
obj.list <- SplitObject(sampleObj, split.by="donor")
geneNMF_programs <- multiNMF(obj.list, k=5)
geneNMF_genes <- getNMFgenes(geneNMF_programs)

}
