% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNMFgenesConsensus}
\alias{getNMFgenesConsensus}
\title{Get consensus gene signature for conserved gene programs}
\usage{
getNMFgenesConsensus(
  nmf.res,
  method = 0.5,
  max.genes = 50,
  hclust.method = "ward.D2",
  nprograms = 10,
  min.confidence = 0.5,
  return.coverage = FALSE,
  plot.tree = TRUE,
  return.tree = FALSE
)
}
\arguments{
\item{nmf.res}{A list of NMF models obtained from `runNMF`}

\item{method}{Parameter passed to `NMF::extractFeatures` to obtain top genes
for each program}

\item{max.genes}{Max number of genes for each programs}

\item{hclust.method}{Method to build similarity tree between individual programs}

\item{nprograms}{Total number of consensus programs}

\item{min.confidence}{Percentage of programs in which a gene is seen, to be retained
in the consensus metaprograms}

\item{return.coverage}{Return sample coverage for each meta-program}

\item{plot.tree}{Whether to plot the similarity tree between gene programs}

\item{return.tree}{Whether to return the similarity tree between gene programs}
}
\value{
Returns a list of top genes for each gene program
}
\description{
Run it over a list of NMF models obtained using `runNMF`; it will determine
gene programs that are consistently observed across samples and values of k.
}
\examples{
# markers <- getNMFgenesConsensus(nmf_results.list)

}
