% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNMFheatmap}
\alias{getNMFheatmap}
\title{Plot heatmap of similarity between NMF programs}
\usage{
getNMFheatmap(
  nmf.res,
  method = 0.5,
  max.genes = 50,
  hclust.method = "ward.D2",
  jaccard.cutoff = c(0, 0.8)
)
}
\arguments{
\item{nmf.res}{A list of NMF models obtained from `multiNMF`}

\item{method}{Parameter passed to `NMF::extractFeatures` to obtain top genes
for each program}

\item{max.genes}{Max number of genes for each programs}

\item{hclust.method}{Method to build similarity tree between individual programs}

\item{jaccard.cutoff}{Min and max values for plotting the Jaccard index}
}
\value{
Returns a heatmap visualization of similarity between gene programs
}
\description{
Run it over a list of NMF models obtained using `multiNMF`; it plots the Jaccard Index
similarity between all pairs of programs calculated over different samples and different
values of k.
}
\examples{
# nmf_genes <- getNMFheatmap(nmf_results.list)

}
