% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multiNMF}
\alias{multiNMF}
\title{RunNMF on a list of Seurat objects}
\usage{
multiNMF(
  obj.list,
  assay = "RNA",
  slot = "data",
  k = 5:6,
  hvg = NULL,
  nfeatures = 2000,
  L1 = c(0, 0),
  min.exp = 0.01,
  max.exp = 3,
  calculate_hvg = TRUE,
  do_centering = TRUE,
  exclude_ribo_mito = FALSE,
  seed = 123
)
}
\arguments{
\item{obj.list}{A list of Seurat objects}

\item{assay}{Get data matrix from this assay}

\item{slot}{Get data matrix from this slot (=layer)}

\item{k}{Number of target components for NMF (can be a vector)}

\item{hvg}{List of pre-calculated variable genes to subset the matrix.
If hvg=NULL and calculate_hvg=NULL, uses all genes.}

\item{nfeatures}{Number of HVG, if calculate_hvg=TRUE}

\item{L1}{L1 regularization term for NMF}

\item{min.exp}{Minimum average log-expression value for retaining genes}

\item{max.exp}{Maximum average log-expression value for retaining genes}

\item{calculate_hvg}{Whether to compute highly variable features}

\item{do_centering}{Whether to center the data matrix}

\item{exclude_ribo_mito}{Exclude ribosomal and mitochondrial genes from
data matrix}

\item{seed}{Random seed}
}
\value{
Returns a list of NMF results
}
\description{
Get the gene expression matrix from a Seurat object, optionally centered
and/or subset on highly variable genes
}
\examples{
# NMF_results <- multiNMF(obj.list)

}
