% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMetaPrograms}
\alias{plotMetaPrograms}
\title{Visualizations for meta-programs}
\usage{
plotMetaPrograms(
  mp.res,
  jaccard.cutoff = c(0, 0.8),
  heatmap.scale = "none",
  heatmap.palette = viridis(100, option = "A", direction = -1),
  heatmap.main = "Clustered Heatmap",
  ...
)
}
\arguments{
\item{mp.res}{The meta-programs object generated by `getMetaPrograms`}

\item{jaccard.cutoff}{Min and max values for plotting the Jaccard index}

\item{heatmap.scale}{Heatmap rescaling (passed to pheatmap as 'scale')}

\item{heatmap.palette}{Heatmap color palette (passed to pheatmap as 'color')}

\item{heatmap.main}{Heatmap title (passed to pheatmap as 'main')}

\item{...}{Additional parameters for pheatmap}
}
\value{
Returns a list with a clustered heatmap and a dendrogram plots of MP similaritites
}
\description{
Generates clustered heatmap and dendrogram for meta-program similarities (Jaccard index).
This function is meant to be run on the object generated by 'GeneNMF::getMetaPrograms', which
contains a pre-calculated tree of pairwise similarities between clusters (as a 'hclust' object).
}
\examples{
# nmf_genes <- plotMetaPrograms(mp.res)

}
