% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{renameMetaPrograms}
\alias{renameMetaPrograms}
\title{Rename meta-programs}
\usage{
renameMetaPrograms(mp.res, old.names = NULL, new.names = NULL)
}
\arguments{
\item{mp.res}{The meta-programs object generated by \code{\link{getMetaPrograms}}}

\item{old.names}{A vector with the current names of the MPs}

\item{new.names}{A vector with the new names of the MPs, align to the order of old.names}
}
\value{
The renamed meta-program object, with MPs ordered according to the order of new.names
}
\description{
Rename meta-programs in the GeneNMF results. This can be useful
to assign more meaningful names to the meta-programs. Or to
rename them after filtering or dropping some of the MPs.
}
\examples{
library(Seurat)
data(sampleObj)
geneNMF_programs <- multiNMF(list(sampleObj), k = 5)
geneNMF_metaprograms <- getMetaPrograms(geneNMF_programs, nMP = 4)
geneNMF_metaprograms_filtered <- dropMetaPrograms(geneNMF_metaprograms, dropMP = c("MP2"))
geneNMF_metaprograms_renamed <- renameMetaPrograms(geneNMF_metaprograms_filtered,
 old.names = c("MP1", "MP3", "MP4"),
 new.names = c("MP1", "MP2", "MP3")
)

}
