% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runGSEA}
\alias{runGSEA}
\title{Run gene-set overrepresentation analysis}
\usage{
runGSEA(
  genes,
  universe = NULL,
  category = "H",
  subcategory = NULL,
  species = "Homo sapiens",
  custom.db = NULL,
  pval.thr = 0.05
)
}
\arguments{
\item{genes}{A vector of genes}

\item{universe}{Background universe of gene symbols
(passed on to \code{fgsea::fora})}

\item{category}{MSigDB main category/collection (e.g. "H" or "C5")}

\item{subcategory}{MSigDB subcategory/subcollection}

\item{species}{Species for overrepresentation analysis. For a list of 
the available species, type \code{msigdbr::msigdbr_species()}}

\item{custom.db}{A custom database of signatures, to be used instead of MSigDB. 
Provide custom.db as a named list, where the list names are the signature 
names, and the list elements are the signatures (as vectors).}

\item{pval.thr}{Min p-value to include results}
}
\value{
Returns a table of enriched gene programs
}
\description{
Utility function to run Gene Set Overrepresentation analysis (ORA) - based 
on \code{fgsea::fora}  against gene sets from MSigDB.
Note: using this function is conditional
to the installation of suggested packages \code{fgsea} and \code{msigdbr}.
}
\examples{
\donttest{
data(sampleObj)
geneset <- c("BANK1","CD22","CD79A","CD19","IGHD","IGHG3","IGHM")
#test is conditional on availability of suggested packages
if (requireNamespace("fgsea", quietly=TRUE) &
   requireNamespace("msigdbr", quietly=TRUE)) {
   gsea_res <- runGSEA(geneset,
       universe=rownames(sampleObj),
       category = "C8")
}
}
}
