% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RunNMF}
\alias{RunNMF}
\title{Compute NMF as a low-dim embedding for Seurat}
\usage{
RunNMF(
  obj,
  assay = "RNA",
  slot = "data",
  k = 10,
  new.reduction = "NMF",
  hvg = NULL,
  nfeatures = 2000,
  L1 = c(0, 0),
  do_centering = TRUE,
  calculate_hvg = TRUE,
  exclude_ribo_mito = FALSE
)
}
\arguments{
\item{obj}{A seurat object}

\item{assay}{Get data matrix from this assay}

\item{slot}{Get data matrix from this slot (=layer)}

\item{k}{Number of components for low-dim representation}

\item{new.reduction}{Name of new dimensionality reduction}

\item{hvg}{List of pre-calculated variable genes to subset the matrix.
If hvg=NULL and calculate_hvg=NULL, uses all genes.}

\item{nfeatures}{Number of HVG, if calculate_hvg=TRUE}

\item{L1}{L1 regularization term for NMF}

\item{do_centering}{Whether to center the data matrix}

\item{calculate_hvg}{Whether to compute highly variable features}

\item{exclude_ribo_mito}{Exclude ribosomal and mitochondrial genes from
data matrix}

\item{seed}{Random seed}
}
\value{
Returns a Seurat object with a new dimensionality reduction (NMF)
}
\description{
compute and load NMF embeddings for a single object, and store them in Seurat data
structure. They can be used as an alternative to PCA for downstream analyses.
}
\examples{
# seurat <- RunNMF(seurat, k=8)
}
