% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMetaPrograms}
\alias{getMetaPrograms}
\title{Extract consensus gene programs (meta-programs)}
\usage{
getMetaPrograms(
  nmf.res,
  method = 0.5,
  max.genes = 50,
  hclust.method = "ward.D2",
  nprograms = 10,
  min.confidence = 0.5,
  remove.empty = TRUE
)
}
\arguments{
\item{nmf.res}{A list of NMF models obtained from `multiNMF`}

\item{method}{Parameter passed to `NMF::extractFeatures` to obtain top genes
for each program}

\item{max.genes}{Max number of genes for each programs}

\item{hclust.method}{Method to build similarity tree between individual programs}

\item{nprograms}{Total number of meta-programs}

\item{min.confidence}{Percentage of programs in which a gene is seen (out of programs in the corresponding program tree branch/cluster), to be retained
in the consensus metaprograms}

\item{remove.empty}{Whether to remove meta-programs with no genes above confidence threshold}
}
\value{
Returns a list with i) 'metaprograms.genes' top genes for each meta-program; ii) 'metaprograms.metrics' dataframe with meta-programs stats: a) freq. of samples where the MP is present, b) average silhouette width, c) mean Jaccard similarity, and d) number of genes in MP; iii) 'programs.jaccard': matrix of Jaccard similarities between meta-programs; iv) 
'programs.tree': hierarchical clustering of meta-programs (hclust tree); v) 'programs.clusters': meta-program assignment to each program
}
\description{
Run it over a list of NMF models obtained using `multiNMF`; it will determine
gene programs that are consistently observed across samples and values of k.
}
\examples{
# markers <- getMetaPrograms(nmf_results.list)

}
