% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNMFgenes}
\alias{getNMFgenes}
\title{Get list of genes for each NMF program}
\usage{
getNMFgenes(nmf.res, method = 0.5, max.genes = 50)
}
\arguments{
\item{nmf.res}{A list of NMF models obtained from `multiNMF`}

\item{method}{Parameter passed to `NMF::extractFeatures` to obtain top genes
for each program}

\item{max.genes}{Max number of genes for each programs}
}
\value{
Returns a list of top genes for each gene program
}
\description{
Run it over a list of NMF models obtained using `multiNMF`
}
\examples{
# nmf_genes <- getNMFgenes(nmf_results.list)

}
