% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMetaPrograms}
\alias{plotMetaPrograms}
\title{Visualizations for meta-programs}
\usage{
plotMetaPrograms(
  mp.res,
  method = 0.5,
  max.genes = 50,
  hclust.method = "ward.D2",
  jaccard.cutoff = c(0, 0.8),
  ...
)
}
\arguments{
\item{mp.res}{The meta-programs object generated by `getMetaPrograms`}

\item{jaccard.cutoff}{Min and max values for plotting the Jaccard index}
}
\value{
Returns a list with a clustered heatmap and a dendrogram plots of metaprograms similaritites
}
\description{
Generates clustered heatmap and dendrogram for meta-program similities (Jaccard distance)
}
\examples{
# nmf_genes <- plotMetaPrograms(mp.res)

}
